ai/requirements.txt
pydantic~=2.8.2
requests~=2.32.3
openai~=1.30.4
pandas~=2.2.2
pytest~=8.2.2
chromadb~=0.5.15
sentence-transformers~=3.0.1
pyyaml~=6.0.1
shortuuid~=1.0.13
uvicorn~=0.30.1
fastapi~=0.111.0
sse-starlette~=2.1.2
torch~=2.3.1
transformers~=4.45.2
httpx~=0.27.2
pytest-asyncio~=0.24.0